+++ plugins/CoreHome/javascripts/materialize-bc.js
@@ -5,6 +5,10 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 (function () {
+    jQuery.htmlPrefilter = function( html ) {
+        return html;
+    };
+
     $(document).ready(function () {
         window.Materialize = window.M;
         $.fn.sideNav = $.fn.sidenav;