+++ plugins/SegmentEditor/tests/Integration/SegmentQueryDecoratorTest.php
@@ -8,6 +8,7 @@
 
 namespace Piwik\Plugins\SegmentEditor\tests\Integration;
 
+use Piwik\ArchiveProcessor\Rules;
 use Piwik\Plugins\SegmentEditor\API;
 use Piwik\Plugins\SegmentEditor\SegmentQueryDecorator;
 use Piwik\Segment;
@@ -36,6 +37,8 @@ class SegmentQueryDecoratorTest extends IntegrationTestCase
         $this->segmentQueryDecorator = self::$fixture->piwikEnvironment->getContainer()->get(
             'Piwik\Plugins\SegmentEditor\SegmentQueryDecorator');
 
+        Rules::setBrowserTriggerArchiving(false);
+
         /** @var API $segmentEditorApi */
         $segmentEditorApi = self::$fixture->piwikEnvironment->getContainer()->get(
             'Piwik\Plugins\SegmentEditor\API');
@@ -47,6 +50,8 @@ class SegmentQueryDecoratorTest extends IntegrationTestCase
         // test that segments w/ auto archive == false are included
         $segmentEditorApi->add('segment 5', 'visitCount<2', 3, $autoArchive = false);
         $segmentEditorApi->add('segment 6', 'countryCode!=fr', 3, $autoArchive = false);
+
+        Rules::setBrowserTriggerArchiving(true);
     }
 
     /**